プラクティス選択プルダウンをChoices.jsに変更
