+++ app/javascript/packs/application.js
@@ -70,6 +70,6 @@ import '../generation-users.js'
 import '../learning-completion-message.js'
 import '../practices-pages.js'
 import '../users-questions.js'
-import '../company-select.js'
+import '../choices-ui.js'
 import '../users-answers.js'
 import '../training-info-toggler.js'
